feat: add negative keyword lists, reports system, enhanced agent service, and Keywords page rebuild
- NegativeKeywordList + NegativeKeywordListItem models with full CRUD
  endpoints (create, delete, add/remove items, apply to campaigns)
- Report model + reports router with SSE-streamed AI narrative generation
  via Claude CLI, saved to DB with token usage tracking
- AgentService expanded: monthly report prompt, date-windowed data
  gathering, change history/impact analysis sections, per-campaign
  detail aggregation
- Keywords.jsx rebuilt with negative keyword management panel,
  negative keyword lists UI, and ad group selector
- Agent.jsx SSE improvements and error handling hardening
- SQLite migration: COALESCE-based unique index on metrics_segmented
  to prevent NULL-duplicate rows
- Seed updates for negative keyword lists
- 745 new test lines (test_agent.py, test_date_boundaries.py)

+++ backend/app/models/keyword.py
@@ -26,7 +26,7 @@ class Keyword(Base):
     cost_micros = Column(BigInteger, default=0)
     conversions = Column(Float, default=0.0)  # Float — Google Ads returns fractional values
     conversion_value_micros = Column(BigInteger, default=0)  # Revenue in micros
-    ctr = Column(Integer, default=0)  # Stored as micros (e.g., 50000 = 5%)
+    ctr = Column(Float, default=0.0)  # Percentage (5.0 = 5%), consistent with MetricDaily
     avg_cpc_micros = Column(BigInteger, default=0)
     cpa_micros = Column(BigInteger, default=0)  # Cost per acquisition
     quality_score = Column(Integer, default=0)  # Google Ads Quality Score 1-10
